refactor(cli): migrate usgs_water tooling onto discovery
Move usgs_water's refresher/writer/curated_ids/emitter/validator to
earthlens.usgs_water.cli, wiring the first stanza_block config role through
discovery (stanza's _STANZA_BLOCK now projects config_table). Add the shared
write_sibling_index to the public toolkit. Relocate usgs_water's CLI tests to
earthlens-ocean; the generic curate/write_stanza command tests stay in core.
Shrink the guard allow-lists (usgs_water; usgs_water import).
